+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [0.10.x]
 
+## [0.10.2] -- 2025-06-30
+
+### Fixed
+
+- `MinimizeDifferenceWithRef` residuals func now returns an always positive value.
+  - Does not make any difference with `OptimisationAlgorithm` such as `LeastSquares` which take the squared residuals. May be a game changer for other algorithms.
+- Flag to handle `SimulationOutput.get(..., elt=elt)` when `elt: Element` is not in `SimulationOutput.elts`.
+  - Fixed some errors at `Objective` creation.
+
 ## [0.10.1] -- 2025-06-06
 
 ### Fixed